ps_mem.py: more robust presentation of command name
* scripts/ps_mem.py: If the exe symlink contains (deleted),
then append [deleted] or [updated] as appropriate.  Also if
the exe link is not present as was the case for the #prelink#
entry mentioned in the script, then try the first command line
parameter, as this might contain the full path to the program.
Note we don't default to using the cmdline as often it just
contains the name and also arbitrary (information) values
are often present.