output an unscaled number for --total
We don't need to worry about alignment for --tofal
so we might as well output the number that's easiest to consume,
and also the most accurate.  The total is really the only
number that needs to be treated differently like this,
so we don't provide a separate --units option to control
the format of the standard output.