Ensure smaps exists before we assume vm_accuracy=2
I've first used this script on GRsecurity enabled kernels
which tweak the kernel to not provide /proc/<PID>/smaps due
to some security concerns.

Reading through the source it seems that code was blindly
presuming there should be precise vm_accounting available
when in reality it falls back to accounting based on figures
from /proc/<PID>/statm which lead to skewed stats for multi
process services such as httpd, spamd.