python-3.6 flush check raises ret_code 120
Since python 3.6, output is sanitized during interpreter exit.
Because Unbuffered() inherits from 'object', not every method from
TextIOBase is implemented and this causes python setting return
code to 120.
This was spotted in our CI and is a common issue with IO wrappers
inheriting from simple 'object'.

(Python maintainer stated that the previous implementation might
have always failed, but it was not sanitized until python-3.6.)